Now ony RAMPa won't link.
