+++ src/util.cpp
@@ -1,5 +1,6 @@
 
 #include "util.h"
+#include "zsys.h"
 #include <sys/stat.h>
 using namespace std;
 
@@ -200,71 +201,7 @@ namespace util
 #endif
 	}
 	
-	long long zc_atoi64(const char *str)
-	{
-		long long val=0;
-		bool neg = false;
-		if(*str == '-')
-		{
-			neg = true;
-			++str;
-		}
-		while(isdigit(*str))
-		{
-			val*=10;
-			
-			val += *str-'0';
-			
-			++str;
-		}
-		
-		return neg ? -val : val;
-	}
-	long long zc_xtoi64(const char *hexstr)
-	{
-		long long val=0;
-		bool neg = false;
-		if(*hexstr == '-')
-		{
-			neg = true;
-			++hexstr;
-		}
-		while(isxdigit(*hexstr))
-		{
-			val<<=4;
-			
-			if(*hexstr<='9')
-				val += *hexstr-'0';
-			else val+= ((*hexstr)|0x20)-'a'+10;
-			
-			++hexstr;
-		}
-		
-		return neg ? -val : val;
-	}
 	
-	int zc_xtoi(const char *hexstr)
-	{
-		int val=0;
-		bool neg = false;
-		if(*hexstr == '-')
-		{
-			neg = true;
-			++hexstr;
-		}
-		while(isxdigit(*hexstr))
-		{
-			val<<=4;
-			
-			if(*hexstr<='9')
-				val += *hexstr-'0';
-			else val+= ((*hexstr)|0x20)-'a'+10;
-			
-			++hexstr;
-		}
-		
-		return neg ? -val : val;
-	}
 	
 	long ffparse2(const char *string) //bounds result safely between -214748.3648 and +214748.3647
 	{
