+++ src/util.h
@@ -26,9 +26,9 @@ namespace util
 	void regulate_path(std::string& buf);
 	bool create_path(const char* path);
 	char* zc_itoa(int value, char* str, int base = 10);
-	long long zc_atoi64(const char *str);
-	long long zc_xtoi64(const char *hexstr);
-	int zc_xtoi(const char *hexstr);
+	//long long zc_atoi64(const char *str);
+	//long long zc_xtoi64(const char *hexstr);
+	//int zc_xtoi(const char *hexstr);
 	long ffparse2(const char *string);
 	long ffparseX(const char *string);
 	int xtoi(char *hexstr);
